Add wiki index checks in CLI.

+++ src/moin/cli/maint/index.py
@@ -1,5 +1,5 @@
 # Copyright: 2011 MoinMoin:MichaelMayorov
-# Copyright: 2023 MoinMoin project
+# Copyright: 2023-2024 MoinMoin:UlrichB
 # License: GNU GPL v2 (or any later version), see LICENSE.txt for details.
 
 """
@@ -14,11 +14,14 @@ from flask.cli import FlaskGroup
 
 from moin.app import create_app, init_backends
 from moin.constants.keys import LATEST_REVS, ALL_REVS
+from moin.utils.filesys import wiki_index_exists
+
 
 from moin import log
 
 logging = log.getLogger(__name__)
 
+ERR_NO_INDEX = "Error: Wiki index does not exist."
 
 @click.group(cls=FlaskGroup, create_app=create_app)
 def cli():
@@ -35,7 +38,6 @@ def cli_IndexCreate(tmp, index_create, storage_create):
         logging.info("options -i or --index-create are obsolete and will be ignored")
     if storage_create:
         logging.info("options -s or --storage-create are obsolete and will be ignored")
-    logging.info("Index creation started")
     return IndexCreate(tmp=tmp)
 
 
@@ -43,17 +45,24 @@ def IndexCreate(**kwargs):
     """
     Create empty indexes
     """
+    if wiki_index_exists():
+        logging.error("Error: wiki index exists. Please check and destroy index before running index-create")
+        return False
     logging.info("Index creation started")
     init_backends(app, create_backend=True)
     tmp = kwargs.get('tmp')
     app.storage.create(tmp=tmp)
     logging.info("Index creation finished")
+    return True
 
 
 @cli.command('index-destroy', help='Destroy the indexes')
 @click.option('--tmp', is_flag=True, required=False, default=False,
               help='use the temporary location.')
 def IndexDestroy(tmp):
+    if not wiki_index_exists():
+        logging.error(ERR_NO_INDEX)
+        raise SystemExit(1)
     logging.info("Index destroy started")
     app.storage.destroy(tmp=tmp)
     logging.info("Index destroy finished")
@@ -68,6 +77,9 @@ def IndexDestroy(tmp):
 @click.option('--index-create', '-i', is_flag=True, required=False, default=False)
 @click.option('--storage-create', '-s', is_flag=True, required=False, default=False)
 def IndexBuild(tmp, procs, limitmb, **kwargs):
+    if not wiki_index_exists():
+        logging.error("{} Run 'moin index-create' first.".format(ERR_NO_INDEX))
+        raise SystemExit(1)
     logging.info("Index build started")
     flaskg.add_lineno_attr = False  # no need to add lineno attributes while building indexes
     app.storage.rebuild(tmp=tmp, procs=procs, limitmb=limitmb)
@@ -77,6 +89,9 @@ def IndexBuild(tmp, procs, limitmb, **kwargs):
 @cli.command('index-update', help='Update the indexes')
 @click.option('--tmp', is_flag=True, required=False, default=False, help='use the temporary location.')
 def IndexUpdate(tmp):
+    if not wiki_index_exists():
+        logging.error(ERR_NO_INDEX)
+        raise SystemExit(1)
     logging.info("Index update started")
     app.storage.update(tmp=tmp)
     logging.info("Index update started")
@@ -99,6 +114,9 @@ def IndexOptimize(tmp):
     """
     Optimize the indexes
     """
+    if not wiki_index_exists():
+        logging.error(ERR_NO_INDEX)
+        raise SystemExit(1)
     logging.info("Index optimization started")
     app.storage.optimize_index(tmp=tmp)
     logging.info("Index optimization finished")
@@ -108,6 +126,9 @@ def IndexOptimize(tmp):
 @click.option('--tmp', is_flag=True, required=False, default=False, help='use the temporary location.')
 @click.option('--truncate/--no-truncate', default=True, help='truncate long entries')
 def IndexDump(tmp, truncate):
+    if not wiki_index_exists():
+        logging.error(ERR_NO_INDEX)
+        raise SystemExit(1)
     logging.info("Index dump started")
     for idx_name in [LATEST_REVS, ALL_REVS]:
         print(" {0} {1} {2}".format("-" * 10, idx_name, "-" * 60))
